✨ Feat. 게임판 및 카드 배치 인터페이스 구현
Why:
게임 플레이 화면에서 카드 배치 및 유저 상호작용을 시각적으로 제공하기 위해
게임판 보드, 카드 선택 UI, 셀 클릭 이벤트 등을 구현할 필요가 있었습니다.

How:
- GameBoardView: 전체 플레이어 리스트와 카드 보드 그리드 UI 구성
- BoardGridView: 9x5 격자 기반 카드 보드 구현 및 셀 탭 처리 추가
- GridCellView: 셀 상태에 따른 이미지/배경/커서 레이어 구성
- CardSelectionView: 선택 가능한 카드 목록과 선택 처리 UI 구현
- BoardViewModel: 카드 배치 로직, 턴 처리, 우승 조건 판단 등 게임 로직 구현

Tag: #feat #gameplay #swiftui #board #viewmodel #multiplayer
